Add files via upload
I applied filters to SQL queries to get specific information on login attempts and employee 
machines. I used two different tables, log_in_attempts and employees. I used the AND, OR, 
and NOT operators to filter for the specific information needed for each task. I also used LIKE 
and the percentage sign (%) wildcard to filter for patterns.